+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to this project will be documented in this file.
 
 [Full diff](https://github.com/sider/runners/compare/0.51.1...HEAD)
 
+- **hadolint** Fix `target` option with glob [#2527](https://github.com/sider/runners/pull/2527)
+
 ## 0.51.1
 
 [Full diff](https://github.com/sider/runners/compare/0.51.0...0.51.1)